Add proper formatting for hours
