+++ app/src/main/java/com/codingpit/pvpcplanner/domain/usecase/date/GetLocalHour.kt
@@ -3,8 +3,6 @@ package com.codingpit.pvpcplanner.domain.usecase.date
 import com.codingpit.pvpcplanner.utils.DateFormatter
 import javax.inject.Inject
 
-class GetLocalHour
-@Inject
-constructor() {
+class GetLocalHour @Inject constructor() {
     operator fun invoke() = DateFormatter.getCurrentHour()
 }
