+++ app/src/main/java/com/codingpit/pvpcplanner/ui/home/HomeScreen.kt
@@ -91,55 +91,3 @@ fun HomeComponents(
         }
     }
 }
-
-@Composable
-fun DateSelector(
-    modifier: Modifier = Modifier,
-    selectedDate: String,
-    currentDate: String,
-    nextDayEnabled: Boolean,
-    onPreviewClicked: () -> Unit,
-    onNextClicked: () -> Unit,
-) {
-    Row(
-        modifier =
-            modifier
-                .padding(16.dp)
-                .background(
-                    MaterialTheme.colorScheme.background,
-                    RoundedCornerShape(8.dp),
-                ),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.Absolute.SpaceBetween,
-    ) {
-        IconButton(modifier = Modifier, onClick = onPreviewClicked) {
-            Icon(imageVector = Icons.AutoMirrored.Default.ArrowBack, contentDescription = "Preview")
-        }
-        val selectedDateText =
-            if (selectedDate == currentDate) {
-                stringResource(R.string.current_date_template).format(selectedDate)
-            } else {
-                selectedDate.toString()
-            }
-        Text(selectedDateText)
-        IconButton(modifier = Modifier, enabled = nextDayEnabled, onClick = onNextClicked) {
-            Icon(
-                imageVector = Icons.AutoMirrored.Default.ArrowForward,
-                contentDescription = "Preview",
-            )
-        }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun DateSelector_Preview() {
-    DateSelector(
-        modifier = Modifier.fillMaxWidth(),
-        selectedDate = "2023-09-01",
-        nextDayEnabled = true,
-        onPreviewClicked = { },
-        onNextClicked = { },
-        currentDate = "2023-09-01",
-    )
-}
